Move chinook package to idiomatic testdata dir
https://github.com/golang/go/issues/14715
